Refactor dataclasses
