+++ spatialstencil/placement/graph.py
@@ -1,113 +1,9 @@
-from dataclasses import dataclass
-from enum import Enum
 from typing import Sequence
 import igraph as ig
 import numpy as np
 
-
-@dataclass
-class FieldDomain:
-    """
-    Represents the 3D domain of a field
-
-    Represented as a 2 by 3 array
-    column 0 -> x, column 1 -> y, column 2 -> z
-    The first row is the lower bound of the domain
-    The second row is the upper bound of the domain
-    """
-    domain: np.ndarray
-
-    def x(self):
-        """
-        Returns the x coordinate upper and lower bounds of the domain
-        :return:
-        """
-        return self.domain[:, 0]
-
-    def y(self):
-        """
-        Returns the y coordinate upper and lower bounds of the domain
-        :return:
-        """
-        return self.domain[:, 1]
-
-    def z(self):
-        """
-        Returns the z coordinate upper and lower bounds of the domain
-        :return:
-        """
-        return self.domain[:, 2]
-
-    # Check if a domain is valid
-    # A domain is valid if it is a 2x3 shaped array and all its elements are integers
-    # column 0 -> x, column 1 -> y, column 2 -> z
-    def __post_init__(self):
-        assert self.domain.shape == (2, 3)
-        assert np.issubdtype(self.domain.dtype, np.integer)
-        # Check that the lower bound is less than or equal to the upper bound
-        assert np.all(self.domain[0] <= self.domain[1])
-
-    def __eq__(self, other):
-        return np.array_equal(self.domain, other.domain)
-
-    def x_length(self):
-        return self.domain[1][0] - self.domain[0][0]
-
-    def y_length(self):
-        return self.domain[1][1] - self.domain[0][1]
-
-    def z_length(self):
-        return self.domain[1][2] - self.domain[0][2]
-
-    def xy_plane_area(self):
-        return self.x_length() * self.y_length()
-
-    def volume(self):
-        return self.z_length() * self.xy_plane_area()
-
-
-
-@dataclass
-class StencilShape:
-    # Represented as a k by 3 array
-    # column 0 -> x, column 1 -> y, column 2 -> z
-    shape: np.ndarray
-
-    # Check if a stencil is valid
-    # A stencil is valid if it is a k times 3 shaped array and all its elements are integers
-    def __post_init__(self):
-        assert self.shape.shape[1] == 3
-        assert self.shape.shape[0] > 0
-        assert np.issubdtype(self.shape.dtype, np.integer)
-
-    def __eq__(self, other):
-        return np.array_equal(self.shape, other.shape)
-
-    def is_horizontal(self) -> bool:
-        """
-        A stencil is horizontal if all z values (3rd coordinate) are 0
-        :return: bool   True if the stencil is horizontal, False otherwise
-        """
-        return np.all(self.shape[:, 2] == 0)
-
-    def is_vertical(self) -> bool:
-        """
-        A stencil is vertical if all x and y values (1st and 2nd coordinates) are 0
-        :return: bool   True if the stencil is vertical, False otherwise
-        """
-        return np.all(self.shape[:, 0:2] == 0)
-
-
-class StencilDirection(Enum):
-    """
-    Indicates the order in which the z-coordinate of the stencil is traversed
-    """
-    # The stencil is traversed in parallel
-    PARALLEL = 0
-    # The stencil is traversed in a forward direction (increasing z)
-    FORWARD = 1
-    # The stencil is traversed in a backward direction (decreasing z)
-    BACKWARD = 2
+from spatialstencil.placement.domain import FieldDomain
+from spatialstencil.placement.stencil import StencilShape, StencilDirection
 
 
 class StencilGraph:
